refactor: FSD 구조 적용하여 Calendar를 Widget 레이어로 승격
