+++ apps/web/src/shared/ui/calendar/index.ts
@@ -1,5 +1,6 @@
-export { Calendar } from './Calendar';
-export type { CalendarProps } from './Calendar';
+export { CalendarGrid } from './CalendarGrid';
+export { CalendarHeader } from './CalendarHeader';
+export { DateCell, type DateCellProps } from './DateCell';
 
 export { formatYearMonth, isSameDate, isToday } from './lib';
 export type { CalendarDate } from './lib';
