+++ apps/web/src/shared/ui/calendar/model/index.ts
@@ -1,3 +1,2 @@
-export { useCalendar, type UseCalendarProps } from './useCalendar';
 export { useWeeklyCarousel } from './useWeeklyCarousel';
 export { WEEKDAYS } from './constants';
