+++ apps/web/src/shared/ui/index.ts
@@ -45,7 +45,6 @@ export { InputField } from './inputField';
 export type { CategoryIconType } from './categoryBtn';
 
 export { Toggle } from './toggle';
-export { Calendar } from './calendar';
 
 export { PageIndicator } from './pageIndicator';
 export type { PageIndicatorProps } from './pageIndicator';
