+++ apps/web/src/widgets/calendar/index.ts
@@ -1,0 +1,1 @@
+export { Calendar, type CalendarProps } from './ui/Calendar';
